Add files via upload
My understanding coding has evolved drastically since this first old project, specifically I'm not so silly as to just be adding in code into my main, but this was a really entertaining project I had a lot of fun creating.

Compile C code, and then when you're GCC'ing it also include the name of the text file you want the system to work with. Code will take text file, determine the probabilistic relationship between word pairings, and then starting with the first word in your text file will randomly create new sentences based on the probabilistic word pair followings. Output set to 1000 words, but can be changed on line 307 of code.